requirements/common.txt
# Common dependencies
flask==2.0.*
python-dotenv==0.19.*